teams 0512

+++ TProject/src/main/java/com/project/controller/TeamsController.java
@@ -1,10 +1,14 @@
 package com.project.controller;
 
+import java.io.PrintWriter;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -52,33 +56,47 @@ public class TeamsController {
 	}
 	
 	@RequestMapping(value="/register.do", method=RequestMethod.GET)
-	public String register() {
+	public String register(HttpServletRequest request) {
+		
+		if (request.getSession().getAttribute("login") == null) {
+			return "redirect:/";
+		}
+		
 		return "teams/register";
 	}
 	
 	@RequestMapping(value="/register.do", method=RequestMethod.POST)
-	public String register(TeamsVO vo, String[] name, int[] capacity) {
+	public String register(HttpServletRequest request, Model model, TeamsVO vo, String[] name, int[] capacity) {
 		
-		List<PartsVO> partsVOlist = new ArrayList<PartsVO>();
+		if (request.getSession().getAttribute("login") == null) {
+
+			model.addAttribute("msg", "로그인이 필요합니다.");
+			model.addAttribute("url", "/member/glogin.do");
+			
+			return "alert";
+			
+		} else {
 		
-		for(int i=0; i<capacity.length; i++) {
-			PartsVO partsVO = new PartsVO();
-			partsVO.setCapacity(capacity[i]);
-			if(name == null) {
-				partsVO.setName("");
-			}else {
-				partsVO.setName(name[i]);
+			List<PartsVO> partsVOlist = new ArrayList<PartsVO>();
+			
+			for(int i=0; i<capacity.length; i++) {
+				PartsVO partsVO = new PartsVO();
+				partsVO.setCapacity(capacity[i]);
+				if(name == null) {
+					partsVO.setName("");
+				}else {
+					partsVO.setName(name[i]);
+				}
+					partsVOlist.add(partsVO);
 			}
-				partsVOlist.add(partsVO);
+			
+			int result = teamsService.register(vo, partsVOlist);
+			if(result != 0) {
+				return "redirect:/teams/main.do";
+			}else {
+				return "teams/register";
+			}
 		}
-		
-		int result = teamsService.register(vo, partsVOlist);
-		if(result != 0) {
-			return "redirect:/teams/main.do";
-		}else {
-			return "teams/register";
-		}
-		
 	}
 	
 	@RequestMapping(value="/details.do")
@@ -93,14 +111,24 @@ public class TeamsController {
 	}
 	
 	@RequestMapping(value="/application.do", method=RequestMethod.GET)
-	public String application(Model model, int teamIdx) {
-		List<PartsVO> partsVO = teamsService.selectParts(teamIdx);
-		TeamsVO teamsVO = teamsService.details(teamIdx);
+	public String application(HttpServletRequest request, Model model, int teamIdx) {
+		if (request.getSession().getAttribute("login") == null) {
+
+			model.addAttribute("msg", "로그인이 필요합니다.");
+			model.addAttribute("url", "/member/glogin.do");
+			
+			return "alert";
+			
+		} else {
 		
-		model.addAttribute("parts", partsVO);
-		model.addAttribute("details", teamsVO);
-		
-		return "teams/application";
+			List<PartsVO> partsVO = teamsService.selectParts(teamIdx);
+			TeamsVO teamsVO = teamsService.details(teamIdx);
+			
+			model.addAttribute("parts", partsVO);
+			model.addAttribute("details", teamsVO);
+			
+			return "teams/application";
+		}
 	}
 	
 	@RequestMapping(value="/application.do", method=RequestMethod.POST)
@@ -112,13 +140,71 @@ public class TeamsController {
 	
 	@RequestMapping(value="/delete.do", method=RequestMethod.POST)
 	@ResponseBody
-	public int delete(int teamIdx) {
+	public String delete(HttpServletRequest request, Model model, int teamIdx) {
+		
+		if (request.getSession().getAttribute("login") == null) {
+
+			model.addAttribute("msg", "로그인이 필요합니다.");
+			model.addAttribute("url", "/member/glogin.do");
+			
+			return "alert";
+			
+		}
+		
 		int result = teamsService.delete(teamIdx);
-		return result;
+		if(result != 0) {
+			return "ok";
+		}
+		return "";
 	}
 	
 	@RequestMapping(value="/update.do", method=RequestMethod.GET)
-	public String update(Model model, int teamIdx) {
+	public String update(HttpServletRequest request, Model model, int teamIdx) {
+		
+		if (request.getSession().getAttribute("login") == null) {
+
+			model.addAttribute("msg", "로그인이 필요합니다.");
+			model.addAttribute("url", "/member/glogin.do");
+			
+			return "alert";
+			
+		}
+		
+		TeamsVO teamsVO = teamsService.details(teamIdx);
+		List<PartsVO> partsVO = teamsService.selectParts(teamIdx);
+		
+		model.addAttribute("details", teamsVO);
+		model.addAttribute("parts", partsVO);
+		
 		return "teams/update";
+	}
+	
+	@RequestMapping(value="/update.do", method=RequestMethod.POST)
+	public String update(HttpServletRequest request, Model model, TeamsVO vo) {
+		
+		if (request.getSession().getAttribute("login") == null) {
+
+			model.addAttribute("msg", "로그인이 필요합니다.");
+			model.addAttribute("url", "/member/glogin.do");
+			
+			return "alert";
+			
+		}
+		
+		int result = teamsService.update(vo);
+		
+		if(result != 0) {
+			return "redirect:/teams/details.do?teamIdx="+vo.getTeamIdx();
+		}else {
+			return "teams/update";
+		}
+	}
+	
+	@RequestMapping(value="/myteams.do")
+	public String myteams(Model model, HttpServletRequest request) {
+		
+		
+		
+		return "teams/myteams";
 	}
 }
